spring-reorganize directories
component 는 app 바깥에!
