+++ crud-frontend/src/app/about/page.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+
+export default function About(){
+
+    return (
+        <div>About</div>
+    )
+}
