+++ crud-frontend/src/app/page.tsx
@@ -3,10 +3,10 @@
 
 import './globals.css';
 import React, {useState, useEffect, useRef} from 'react';
-import Snackbar from './components/Snackbar';
-import Form from './components/Form';
-import List from './components/List';
-import MyComponent from './components/FetchDataComponent';
+import Snackbar from '@/components/Snackbar';
+import Form from '@/components/Form';
+import List from '@/components/List';
+import MyComponent from '../components/FetchDataComponent';
 
 export interface DiaryEntry {
     id: number;
